Fixed html validation: bad value menu for attribute role on element section.
This is in the footer actions portlet, which contains links to
Sitemap, Accessibility, and Contact.

Also, when the header is not shown, you get this warning:
Section lacks heading.

Also, when the header *is* shown, we get an error:
The element 'header' must not appear as a descendant of the 'footer' element.

Also, the section with class portletContent is always lacking a header.

Solutions in this fix:
- Removed the role from this section.
- Changed all sections to divs.