Removed role 'section' from section.
A section is already a section, so there is no need for a role with this name/
Also, no such role exists.
If no role is given explicitly, the default 'region' role is used,
and such a default role should not be set, as browsers do that themselves.

See http://w3c.github.io/html/sections.html#the-section-element