Downgraded latest changelog entry from breaking change to new feature.
According to semver.org, a newer dependency does not constitute a breaking change.
Here it does mean that you need Plone 5.2 and it will not work with 5.1, which also is more on the 'new feature' level.
The code of this package doesn't really change, only a test.
For a breaking change, we would have to update the major version.
It's a grey area.

[ci skip]